requirements.txt
# Telegram Bot
python-telegram-bot==21.0

# Exchange APIs
ccxt==4.2.25

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# Environment & Config
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2